requirements.txt
torch
torchvision
matplotlib
flask
pillow
